refactor: Controller에서만 사용하는 Logger 패키지 수정
