+++ src/main/java/controller/Logger.java
@@ -1,4 +1,4 @@
-package common;
+package controller;
 
 public class Logger {
